# Time_Management_Project/services/back-main/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Time Management Project",
    "description": "Website for tracking the working hours of employees of an IT company",
    "version": "1.0.0"
  },
  "host": "localhost:9000",
...
  "schemes": "http",
  "paths": {
    "/developers/get": {
      "get": {
        "tags": ["Developers"],

        "summary": "Returns a developer.",
        "description": "Endpoint to get a developer.",
        "parameters": [
           {
            "name": "page",
            "in": "query",
            "required": true,
            "type": "integer"
          },
          {
            "name": "pagSize",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
                 }
                     }
              }
                },
    "/developers/get/{id}": {
      "get": {
        "tags": ["Developers"],
        "summary": "Returns a developer by id.",
        "description": "Endpoint to get a specific developer by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
           {
            "name": "page",
            "in": "query",
            "required": true,
            "type": "integer"
          },
          {
            "name": "pagSize",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
                 }
                     }
            }
                  },                
    "/developers/get_by_task/{id_task}": {
      "get": {
        "tags": ["Developers"],
        "summary": "Returns a developers by task.",
        "description": "Endpoint to get a developers by task.",
        "parameters": [
          {
            "name": "id_task",
            "in": "path",
            "required": true,
            "type": "integer"
          },
           {
            "name": "page",
            "in": "query",
            "required": true,
            "type": "integer"
          },
          {
            "name": "pagSize",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/developers/get_by_project/{id_project}": {
      "get": {
        "tags": ["Developers"],
        "summary": "Returns a developers by project.",
        "description": "Endpoint to get a developers by project.",
        "parameters": [
          {
            "name": "id_project",
            "in": "path",
            "required": true,
            "type": "integer"
          },
           {
            "name": "page",
            "in": "query",
            "required": true,
            "type": "integer"
          },
          {
            "name": "pagSize",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/developers/update/{id}": {
      "post": {
        "tags": ["Developers"],
        "summary": "Updating information about the developer.",
        "description": "Endpoint to update information about the developer. name: string, surname: string, patronymic: string, birth: date, email: string",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/developers/put-avatar/{id}": {
      "post": {
        "tags": ["Developers"],
        "summary": "Put avatar.",
        "description": "Endpoint put avatar.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "file",
            "in": "formData",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
"/developers/delete/{id}": {
      "get": {
        "tags": ["Developers"],
        "summary": "Delete developer by id.",
        "description": "Endpoint to delete a developers by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/developers/count_by_task/{task_id}": {
      "get": {
        "tags": ["Developers"],
        "summary": "Count by task.",
        "description": "Endpoint count by task.",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/developers/count_by_project/{project_id}": {
      "get": {
        "tags": ["Developers"],
        "summary": "Count by project.",
        "description": "Endpoint count by project.",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/developers/put": {
      "post": {
        "tags": ["Developers"],
        "summary": "Put developer.",
        "description": "name: string; surname: string; patronymic: string; birth: date; email: string, position: integer; specialty: integer,isAdmin: boolean; password: string;",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}